refactor(frontend): move categories to home folder
