+++ packages/code-du-travail-frontend/src/home/__tests__/Categories.test.js
