+++ packages/code-du-travail-ui/src/Categories.js
@@ -1,9 +1,14 @@
 import React from "react";
+import PropTypes from "prop-types";
 
-const Categories = props => (
+const Categories = ({ className = "", ...props }) => (
   <div className="categories">
-    <ul className="categories__list" {...props} />
+    <ul className={`categories__list ${className}`} {...props} />
   </div>
 );
+
+Categories.propTypes = {
+  className: PropTypes.string
+};
 
 export default Categories;
