+++ packages/code-du-travail-ui/src/Category.js
@@ -1,28 +1,21 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-const Category = ({ href, icon, title, text, ...props }) => (
-  <li className="categories__list-item" {...props}>
-    <a href={href}>
-      <img src={icon} alt={title} />
-      <h3>{title}</h3>
-      <p>{text}</p>
-    </a>
-  </li>
+const Category = ({ small, ...props }) => (
+  <li
+    className={`categories__list-item${
+      small ? " categories__list-item--small" : ""
+    }`}
+    {...props}
+  />
 );
 
 Category.propTypes = {
-  href: PropTypes.string.isRequired,
-  icon: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  text: PropTypes.string.isRequired
+  small: PropTypes.bool
 };
 
 Category.defaultProps = {
-  href: "#",
-  icon: "assets/icons/handshake.svg",
-  title: "Embauche et contrat",
-  text: "Fluctuat nec mergitur."
+  small: false
 };
 
 export default Category;
